+++ src/xrSound/SoundRender_Emitter_FSM.cpp
@@ -152,7 +152,7 @@ void CSoundRender_Emitter::update(float dt)
 	case stStartingLooped:
 		if (iPaused) break;
 		fTimeStarted = fTime;
-		fTimeToStop = 0xffffffff;
+		fTimeToStop = (float)0xffffffff;
 		fTimeToPropagade = fTime;
 		fade_volume = 1.f;
 		occluder_volume = SoundRender->get_occlusion(p_source.position, .2f, occluder);
